random: fix typo
When seeding with a single u64 word, we don't want to assign all 16
values in the state array with the same seed word, so we add a value
to it to make it slightly different. However, the old code added 1 to
every seed, making all 16 words identical; this was due to a typo.